src/main/java/com/ninjaone/dundie_awards/controller/rest/EmployeeController.java
package com.ninjaone.dundie_awards.controller.rest;

import static org.springframework.http.HttpStatus.CREATED;
import static org.springframework.http.HttpStatus.NO_CONTENT;

import java.util.List;

import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import com.ninjaone.dundie_awards.dto.EmployeeDto;
import com.ninjaone.dundie_awards.dto.EmployeeUpdateRequestDto;
import com.ninjaone.dundie_awards.service.EmployeeService;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("/employees")
@RequiredArgsConstructor
public class EmployeeController {

	private final EmployeeService employeeService;

	@GetMapping
	public List<EmployeeDto> getAllEmployees() {
		return employeeService.getAllEmployees();
	}
	
	@GetMapping("/{id}")
	public EmployeeDto getEmployeeById(@PathVariable("id") long id) {
		return employeeService.getEmployee(id);
	}

	@PostMapping
	@ResponseStatus(CREATED)
	public EmployeeDto createEmployee(@Valid @RequestBody EmployeeDto employee) {
		return employeeService.createEmployee(employee);
	}

	@PutMapping("/{id}")
	public EmployeeDto updateEmployee(@PathVariable("id") long id, @Valid @RequestBody EmployeeUpdateRequestDto updateRequest) {
		return employeeService.updateEmployee(id, updateRequest);
	}

	@DeleteMapping("/{id}")
	@ResponseStatus(NO_CONTENT)
	public void deleteEmployee(@PathVariable("id") long id) {
		employeeService.deleteEmployee(id);
	}
}
